React to PR comments
