+++ src/SamplesApp/UITests.Shared/Windows_UI_Xaml/UIElementTests/TransformToVisual_ScrollViewer.xaml.cs
@@ -5,7 +5,7 @@ using Windows.Foundation;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Uno.UI.Samples.Controls;
-
+using Uno.UI.Samples.UITests.Helpers;
 
 namespace UITests.Shared.Windows_UI_Xaml.UIElementTests
 {
